+++ tnnt_templates/context_processors.py
@@ -23,45 +23,13 @@ def tnnt_settings(request: WSGIRequest) -> dict:  # pylint: disable=unused-argum
         "REGISTRATION_VERIFY_EMAIL": getattr(
             settings, "REGISTRATION_VERIFY_EMAIL", True
         ),
+        "TNNT_TEMPLATE_ENTITY_TYPE": getattr(
+            AppSettings, "TNNT_TEMPLATE_ENTITY_TYPE", "alliance"
+        ),
+        "TNNT_TEMPLATE_ENTITY_ID": getattr(AppSettings, "TNNT_TEMPLATE_ENTITY_ID", 1),
+        "TNNT_TEMPLATE_ENTITY_NAME": getattr(
+            AppSettings, "TNNT_TEMPLATE_ENTITY_NAME", ""
+        ),
     }
-
-    # entity ID
-    # 1 if none is given
-    try:
-        return_value["TNNT_TEMPLATE_ENTITY_ID"] = AppSettings.TNNT_TEMPLATE_ENTITY_ID
-    except AttributeError:
-        return_value["TNNT_TEMPLATE_ENTITY_ID"] = 1
-
-    # entity type
-    # 'alliance' if none is given
-    try:
-        return_value["TNNT_TEMPLATE_ENTITY_TYPE"] = (
-            AppSettings.TNNT_TEMPLATE_ENTITY_TYPE
-        )
-    except AttributeError:
-        return_value["TNNT_TEMPLATE_ENTITY_TYPE"] = "alliance"
-
-    # entity name
-    # empty if none is given
-    try:
-        return_value["TNNT_TEMPLATE_ENTITY_NAME"] = (
-            AppSettings.TNNT_TEMPLATE_ENTITY_NAME
-        )
-    except AttributeError:
-        return_value["TNNT_TEMPLATE_ENTITY_NAME"] = ""
-
-    try:
-        return_value["TNNT_TEMPLATE_URLS_OWN_WEBSITES"] = (
-            AppSettings.TNNT_TEMPLATE_URLS_OWN_WEBSITES
-        )
-    except AttributeError:
-        pass
-
-    try:
-        return_value["TNNT_TEMPLATE_URLS_OTHER_WEBSITES"] = (
-            AppSettings.TNNT_TEMPLATE_URLS_OTHER_WEBSITES
-        )
-    except AttributeError:
-        pass
 
     return return_value